Tidued up variable conversions for gfortran-friendliness.
git-svn-id: https://scm.physics.ox.ac.uk/svn/radtrancode/trunk@348 4dc0e162-f54d-4b9e-8a18-81c7a7f2ddce

+++ nemesis/forwardnogMCS.f
@@ -320,7 +320,7 @@ C       listed, assume that channels are in usual order
          vv = vconv(igeom,i)
          if (vv.lt.350.)then
 C         Get B pixel numbers
-          ipixB = 100*(0.001+vv*10-int(vv*10))
+          ipixB = int(100*(0.001+vv*10-int(vv*10)))
           goto 646
          endif
         enddo
